MB-25475: reorganize Analytics modules
In order to prepare it to lazy loading.

This change introduces new main.js module which
is a root module of Analytics UI and place for
Analytics UI inital config.

In this change, we also indicate that the main.js
will be imported into generated pluggable-uis.js
(more here http://review.couchbase.org/c/ns_server/+/125551)

Change-Id: Ia742880da528853d107a75b1fd43ec036e0dff7c
Reviewed-on: http://review.couchbase.org/c/cbas-ui/+/125692
Tested-by: Pavel Blagodov <stochmail@gmail.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Pavel Blagodov <stochmail@gmail.com>