Fix several minor UI issues.
Query formatting was not working due to a typo.

Query and result sizing was not occurring properly, for several
reasons. 1) there was an extra call made to updateEditorSizes from
code in the HTML, which was inappropriate and unnecessary. 2) We
should have been using _.throttle instead of _.debounce.

A warning was being incorrectly issued about the data being too large
for the table viewer, when in fact the table viewer can show 10x the
data as the JSON view.

Change-Id: I784668d622ed9eab475acf1c1f55971d262ceae3
Reviewed-on: http://review.couchbase.org/c/cbas-ui/+/126064
Reviewed-by: Murtadha Hubail <Murtadha.hubail@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>