MB-36296 - support for collections in Analytics Workbench UI
The analytics workbench UI needed to allow the user to specify a
bucket/scope/collection for datasets associated with local and remote
Couchbase links. In the case of remote Couchbase links, needed to make
scope/collection optional in case the remote cluster is pre-7.0. In
the case of the local link, use the QwCollectionMenu component from
query-ui, for remote links we have text fields.

Change-Id: Ieafbc96ab51c71162e1323233c00d703d1905bfc
Reviewed-on: http://review.couchbase.org/c/cbas-ui/+/142650
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>