MB-44633 - link and dataset not shown when backticks present
Dataverses have a real name and a display name which may be
different. The bug was that the HTML was searching for links and
datasets by display name, thus not finding them when the real name and
display name differed.

Also begin the process of changing nomenclature from "dataverses" and
"datasets" to "scopes" and "collections".

Change-Id: I3234f416524bb99cb68ef3e87b53902c7daed4ff
Reviewed-on: http://review.couchbase.org/c/cbas-ui/+/147521
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>