MB-44768 - Error dialog showing too much information
With an error response from the server, the error dialog was showing
the entire contents of the data field of the HTTP response, instead of
just showing data.errors, the list of error messages. Added a 2 line
check for data.errors, showing it if present.

Change-Id: I69a97e902706222524266726b5ef61cbbf5a8f69
Reviewed-on: http://review.couchbase.org/c/cbas-ui/+/148069
Tested-by: Eben Haber <eben@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>