MB-45754 - Also allow tab-completion of full paths
MB-45854 - Make sure all external links displayed as such.

Before the code tested for the "s3" type to indicate external
links. Now the logic is flipped, so that internal is either no
LinkType (for the Local link), or COUCHBASE (for remote Couchabse
links). Everything else will be treated as external.

Change-Id: I57bf12e312a6fc606c425692515eba19b515e569
Reviewed-on: http://review.couchbase.org/c/cbas-ui/+/151923
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>