MB-45701 - need to quote both two-part and one-part scope names
MB-45886 - Datasets in one dataverse with link from another not shown

Also fix some problems introduced due to a bad merge of an earlier
change.

Finally, improve the metadata processing code. The previous code was
complex because it tried to handle the dataverses, links, and datasets
in any order. (E.g., handling cases such as seeing a dataset before
the dataverse that it is part of.) This change includes a modification
to the metadata query to sort the results, so that dataverses come
first, then links, then datasets. That makes the metadata processing
easier and less error prone.

Change-Id: I7c278099c1eda305c9c74d2528ad23ab79eb54cb
Reviewed-on: http://review.couchbase.org/c/cbas-ui/+/152032
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>