MB-45909 - Differentiate display names from names for queries
Analytics needs 3 different forms of dataverse names. 1) The raw name,
used in the REST API, where multipart names are separated by '/'. 2)
The parser name, where '/' is not used, and multipart names are
separated by '.' instead (with any identifiers containing special
characters enclosed in back ticks). 3) The name shown to the user in
the UI. Prior to this change we were only maintaining 2 names, now we
are maintaining all 3.

Change-Id: I6fabe40676b6b58778550ea8e37f8480a6402460
Reviewed-on: http://review.couchbase.org/c/cbas-ui/+/152391
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>