MB-46399 - password shown in error dialog after REST failure
This problem occurred when creating a remote link in a mixed
cluster. The REST API url changed between 6.6 and 7.0, and the 7.0
code was using the incorrect URL and getting a 404 error. Since a 404
wasn't providing any error message, the error dialog was showing the
JSON for the entire HTTP response. The solution for this problem is
twofold:

1) Make sure never to display the entire HTTP response in the error
dialog, and

2) Use the correct REST API in mixed clusters

This fix includes both, avoiding the wrong REST APIs in mixed clusters
and improving the error handling to better (and more consistently)
display error messages.

Change-Id: I53a8a18a348c3969d3a3dfecef82eac37f8ac8a2
Reviewed-on: http://review.couchbase.org/c/cbas-ui/+/154076
Tested-by: Michael Blow <michael.blow@couchbase.com>
Reviewed-by: Michael Blow <michael.blow@couchbase.com>