MB-45709 - Use correct names for user display
With the new analytics scopes, dataverses can have 3 distinct names: a
user visible name without too much punctuation, a canonical name for
REST calls, and a name for issuing queries (with appropriate
back-ticks to handle special characters).

When creating links, the dialog box needed to use the display name,
but it also needs to pass the canonical name to the function that
makes the REST call to create the link.

Change-Id: I41389bafc875187900faac8b7074357275740015
Reviewed-on: http://review.couchbase.org/c/cbas-ui/+/158445
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Murtadha Hubail <Murtadha.hubail@couchbase.com>