MB-46565 - UI not always auto-refreshing metadata
The problem here appears to be that the check for whether to refresh
the metadata looked for certain strings in the query, such as "CREATE
EXTERNAL DATASET". However, the check failed if there were extra
spaces in the query string, e.g., "CREATE EXTERNAL". Once I added a
regex substitution to remove multiple spaces in a row, the UI started
behaving correctly.

Change-Id: I711360958a1bafbab79ff8cd659ca7c4e26d0373
Reviewed-on: http://review.couchbase.org/c/cbas-ui/+/158453
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Murtadha Hubail <Murtadha.hubail@couchbase.com>