MB-47841 - make chart visible after navigation
Bug: run a query, create a chart, navigate to Indexes page and
back. Chart is gone, though tabular and JSON views still show the
query result.

The query workbench in AngularJS, while the chart component is Angular
8. The best way for the workbench to inform the chart component about
changes is through an Observable, which it does at the end of each
query execution. However, this didn't account for navagition to other
parts of the UI, since the Component is recreated each time the user
navigates there. Add a line to the controller "activate" function to
update the Observable with any available data when the Component is
created.

Change-Id: I32fa8fd0d5557aa2c88c04626155e01a7cbf236d
Reviewed-on: http://review.couchbase.org/c/cbas-ui/+/159172
Tested-by: Eben Haber <eben@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Murtadha Hubail <Murtadha.hubail@couchbase.com>