MB-46975 - Analytics workbench forgetting the query context
The problem here was that the query context was being stored in the
workbench controller, which is recreated every time you navigate away
and back. Thus it could easily forget the context.

As a solution, I have made the query context part of the query
history. That way you can re-run a query from your history and it will
have the same context as before. You can also have different queries
in the history with different query contexts.

Of course, the first time you load this new version, there will be no
context for anything in the history, but once you start specifying
context and executing commands, they should persist along with the
rest of the history.

Change-Id: Ibad26e87869f7d0d41e725891e0f4ae3418908fe
Reviewed-on: http://review.couchbase.org/c/cbas-ui/+/159256
Reviewed-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Murtadha Hubail <Murtadha.hubail@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>