Update cbas-ui code to work with Angular 8 upgraded query-ui
The cbas-ui uses several services and components from the query-ui. Now
that the query-ui has been upgraded to Angular 8, those services and
components have been upgraded as well. Angular provides "downgrade"
functionality so that cbas-ui, which is in AngularJS, can still refer
to and use the upgraded components and services.

Change-Id: Ia0ba04d1ee2c7a28ff176f9c8d9a010884e962d3
Reviewed-on: http://review.couchbase.org/c/cbas-ui/+/164110
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>