MB-50949 Fix inappropriate result size warning in analytics workbench
The analytics workbench had code to issue a warning about large query
result sizes - originally rendering large results (> 1MB for trees and
tables) could cause the browser to freeze up for tens to hundreds of
seconds. The warning code hasn't been updated, however, since the
table viewer was upgraded (it can now handle results > 100MB) and the
tree viewer was removed (and replaced with charts).

A good way to see the old problem is to map beer-sample, then run:

  select * from `beer-sample`._default._default;

Previously, a warning message would appear for table view
(incorrectly), and would also appear for the chart view (with the
message mentioning the now removed tree view).

This change removed any warning except for the JSON view, which
starts to slow down with result sizes > 10MB.

Change-Id: Ia60e9c9ae716b7da1b5af4bbd794e0a2df069f17
Reviewed-on: https://review.couchbase.org/c/cbas-ui/+/170613
Tested-by: Eben Haber <eben@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Hussain Towaileb <Hussain.Towaileb@Couchbase.com>