MB-52537 - missing CSS style when analytics loaded first
The history dialog for both query and analytics relies on a style
defined in query.css, thus it doesn't get loaded if analytics is
loaded before query. The best solution would be to move the style to
the core CSS in the ns_server code, but given the short time frame
left for 7.1.2, an acceptable solution is to copy the style to
cw_history_dialog.html (as a stop-gap, knowing that all the analytics
UI code will need to be migrated to Angular or Svelte at some point in
the next year).

Change-Id: Ide8976234b43803301b411c10dc5f4499fbb0d16
Reviewed-on: https://review.couchbase.org/c/cbas-ui/+/176530
Tested-by: Eben Haber <eben@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Hussain Towaileb <Hussain.Towaileb@Couchbase.com>
Well-Formed: Restriction Checker