MB-59418 Fix Query with Too Big Response
On Analytics page, if a query result is too big
to be processed in the browser, the UI is stuck until the POST
request finishes.
This fix checks progressively if the output
gets too long, by using progress event handler.
When the intermediate loaded output is bigger than
the set limit (250MB), the query is canceled and
the following message is displayed to the user,
without the "Show Anyway" option:
"The statement results are larger than the
 workbench limit of 250MB and cannot be displayed.
 Please use another client product to access them."
Fixes CBSE CBSE-15275.

Change-Id: I541ed0284ec5b29b44f1f0de20f9564ea08a6f3e
Reviewed-on: https://review.couchbase.org/c/cbas-ui/+/199749
Tested-by: Raluca Lupu <raluca.lupu@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>