MB-59971 - remove _uistats endpoint from cbas-ui
As luck would have it, the cbas-ui hasn't actually used the _uistats
endpoint for a long time, it has been using the stats/range
instead. There was an obsolete function referring to the _uistats
endpoint, but that function wasn't used anywhere. The change removes
that function, and all the stats are still showing up just fine (on
the analytics monitoring page).

Change-Id: Ic1a606616e2d34093e930b7347a597e09b5096fd
Reviewed-on: https://review.couchbase.org/c/cbas-ui/+/202334
Reviewed-by: Till Westmann <till@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>