MB-65767 - fix handling of numbers > 2^53 in the analytics workbench
The code for handling very large numbers in the analytics workbench
had a bug causing no output with query results containing very large
integers. This bug was shared with the query workbench. The query
workbench issue (a regex problem) was fixed some time ago, but the fix
was not added to the analytics workbench.

This fix copies the working regex from query-ui to cbas-ui.

Change-Id: Id64f2eb7d81dfdce1cb1df89aef34f90dae58c6c
Reviewed-on: https://review.couchbase.org/c/cbas-ui/+/224930
Reviewed-by: Peeyush Gupta <peeyush.gupta@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>